General Cleanup

+++ OpenRA.Game/FileFormats/XccLocalDatabase.cs
@@ -49,11 +49,11 @@ namespace OpenRA.FileFormats
 				writer.Write(Encoding.ASCII.GetBytes("XCC by Olaf van der Spek"));
 				writer.Write(new byte[] { 0x1A, 0x04, 0x17, 0x27, 0x10, 0x19, 0x80, 0x00 });
 
-				writer.Write((int)(Entries.Aggregate(Entries.Length, (a, b) => a + b.Length) + 52)); // Size
-				writer.Write((int)0); // Type
-				writer.Write((int)0); // Version
-				writer.Write((int)0); // Game/Format (0 == TD)
-				writer.Write((int)Entries.Length); // Entries
+				writer.Write(Entries.Aggregate(Entries.Length, (a, b) => a + b.Length) + 52); // Size
+				writer.Write(0); // Type
+				writer.Write(0); // Version
+				writer.Write(0); // Game/Format (0 == TD)
+				writer.Write(Entries.Length); // Entries
 				foreach (var e in Entries)
 				{
 					writer.Write(Encoding.ASCII.GetBytes(e));
